Integration testing, load of fixes, and better macro.
- Macro type must be specified.
- The macro now has a special case for string types.
- Maps in the `test_maps` folder now get parsed during testing.
- Many values added in later versions are now optional.
- Fixed rotation renames being swapped.

+++ src/difficulty/gameplay_event.rs
@@ -16,5 +16,5 @@ pub struct BpmEvent {
     #[serde(rename = "b")]
     pub beat: f32,
     #[serde(rename = "m")]
-    pub bpm: i32,
+    pub bpm: f32,
 }
